cardlesscash-queryfix - error handling

+++ src/main/java/com/mashreq/transfercoreservice/cardlesscash/dto/request/CardLessCashQueryRequest.java
@@ -1,7 +1,5 @@
 package com.mashreq.transfercoreservice.cardlesscash.dto.request;
 
-import java.math.BigInteger;
-
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -17,5 +15,5 @@ import lombok.NoArgsConstructor;
 public class CardLessCashQueryRequest {
 
     private String accountNumber;
-    private BigInteger remitNumDays;
+    private Integer remitNumDays;
 }
