+++ src/test/java/com/mashreq/transfercoreservice/cardlesscash/controller/CardLessCashControllerTest.java
@@ -1,17 +1,6 @@
 package com.mashreq.transfercoreservice.cardlesscash.controller;
 
-import com.mashreq.transfercoreservice.cardlesscash.dto.request.CardLessCashBlockRequest;
-import com.mashreq.transfercoreservice.cardlesscash.dto.response.CardLessCashBlockResponse;
-import com.mashreq.transfercoreservice.cardlesscash.service.CardLessCashService;
-import com.mashreq.transfercoreservice.cardlesscash.dto.request.CardLessCashGenerationRequest;
-import com.mashreq.transfercoreservice.cardlesscash.dto.request.CardLessCashQueryRequest;
-import com.mashreq.transfercoreservice.cardlesscash.dto.response.CardLessCashGenerationResponse;
-import com.mashreq.transfercoreservice.cardlesscash.dto.response.CardLessCashQueryResponse;
-import com.mashreq.transfercoreservice.cardlesscash.controller.CardLessCashController;
-import com.mashreq.webcore.dto.response.Response;
-
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -24,6 +13,15 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
+
+import com.mashreq.transfercoreservice.cardlesscash.dto.request.CardLessCashBlockRequest;
+import com.mashreq.transfercoreservice.cardlesscash.dto.request.CardLessCashGenerationRequest;
+import com.mashreq.transfercoreservice.cardlesscash.dto.request.CardLessCashQueryRequest;
+import com.mashreq.transfercoreservice.cardlesscash.dto.response.CardLessCashBlockResponse;
+import com.mashreq.transfercoreservice.cardlesscash.dto.response.CardLessCashGenerationResponse;
+import com.mashreq.transfercoreservice.cardlesscash.dto.response.CardLessCashQueryResponse;
+import com.mashreq.transfercoreservice.cardlesscash.service.CardLessCashService;
+import com.mashreq.webcore.dto.response.Response;
 
 @RunWith(MockitoJUnitRunner.class)
 public class CardLessCashControllerTest {
@@ -89,7 +87,7 @@ public class CardLessCashControllerTest {
     public void queryCardLessCashRequestSuccessTest() {
 
         String accountNumber = "019100064328";
-        BigInteger remitNumDays = BigInteger.valueOf(1);
+        Integer remitNumDays = 1;
         CardLessCashQueryRequest cardLessCashQueryRequest = CardLessCashQueryRequest.builder()
                 .accountNumber(accountNumber)
                 .remitNumDays(remitNumDays)
